+++ apps/backoffice-v2/src/domains/storage/query-keys.ts
@@ -1,9 +1,16 @@
 import { createQueryKeys } from '@lukemorales/query-key-factory';
 import { fetchFileById } from './fetchers';
+import { env } from '@/common/env/env';
 
 export const storageQueryKeys = createQueryKeys('storage', {
-  fileById: (fileId: string) => ({
-    queryKey: [{ fileId }],
-    queryFn: () => fetchFileById(fileId),
+  fileById: ({
+    fileId,
+    withSignedUrl = env.VITE_FETCH_SIGNED_URL,
+  }: {
+    fileId: string;
+    withSignedUrl?: boolean;
+  }) => ({
+    queryKey: [{ fileId, withSignedUrl }],
+    queryFn: () => fetchFileById({ fileId, withSignedUrl }),
   }),
 });