+++ packages/workflow-core/src/lib/utils/definition-validator/definition-validator.ts
@@ -1,0 +1,31 @@
+import { StateMachine } from 'xstate';
+
+import { WorkflowExtensions } from '../../types';
+import { statesValidator, validateTransitionOnEvent } from './states-validator';
+import { extensionsValidator } from './extensions-validator';
+import { AnyRecord } from '@ballerine/common';
+
+export const definitionValidator = (
+  definition: {
+    states?: StateMachine<any, any, any>['states'];
+    initial?: string;
+    initialContext?: AnyRecord;
+  },
+  extensions?: WorkflowExtensions,
+) => {
+  if (!definition.states) return;
+
+  if (definition.initial) {
+    validateTransitionOnEvent({
+      stateNames: Object.keys(definition.states),
+      currentState: 'NULL_AS_UNINITIATED_STATE',
+      targetState: definition.initial,
+    });
+  }
+
+  statesValidator(definition.states);
+
+  if (extensions) {
+    extensionsValidator(extensions, definition.states);
+  }
+};