+++ services/workflows-service/src/auth/auth.service.ts
@@ -1,7 +1,6 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { UserInfo } from '../user/user-info';
 import { UserService } from '../user/user.service';
-import { LoginDto } from './dtos/login';
 import { PasswordService } from './password/password.service';
 
 @Injectable()