+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -63,9 +63,10 @@ import {
   WorkflowRuntimeListQueryResult,
 } from './types';
 import { addPropertiesSchemaToDocument } from './utils/add-properties-schema-to-document';
-import { WorkflowDefinitionRepository } from './workflow-definition.repository';
+import { WorkflowDefinitionRepository } from '../workflow-defintion/workflow-definition.repository';
 import { WorkflowEventEmitterService } from './workflow-event-emitter.service';
 import { WorkflowRuntimeDataRepository } from './workflow-runtime-data.repository';
+import mime from 'mime';
 
 type TEntityId = string;
 
@@ -257,7 +258,10 @@ export class WorkflowService {
         ...workflow.context,
         documents: workflow.context?.documents?.map(
           (document: DefaultContextSchema['documents'][number]) => {
-            return addPropertiesSchemaToDocument(document);
+            return addPropertiesSchemaToDocument(
+              document,
+              workflow.workflowDefinition.documentsSchema,
+            );
           },
         ),
       },
@@ -804,8 +808,8 @@ export class WorkflowService {
       id: documentId,
     };
 
-    const newDocument = addPropertiesSchemaToDocument(document);
-    const propertiesSchema = newDocument?.propertiesSchema ?? {};
+    const documentSchema = addPropertiesSchemaToDocument(document, workflowDef.documentsSchema);
+    const propertiesSchema = documentSchema?.propertiesSchema ?? {};
     if (Object.keys(propertiesSchema)?.length) {
       let propertiesSchemaForValidation = propertiesSchema;
       if (!checkRequiredFields) {
@@ -814,7 +818,7 @@ export class WorkflowService {
       }
       const validatePropertiesSchema = ajv.compile(propertiesSchemaForValidation);
 
-      const isValidPropertiesSchema = validatePropertiesSchema(newDocument?.properties);
+      const isValidPropertiesSchema = validatePropertiesSchema(documentSchema?.properties);
 
       if (!isValidPropertiesSchema) {
         throw new BadRequestException(
@@ -830,7 +834,7 @@ export class WorkflowService {
     const updatedWorkflow = await this.updateContextById(
       workflowId,
       {
-        documents: [newDocument],
+        documents: [documentSchema],
       },
       [projectId],
     );
@@ -1590,15 +1594,33 @@ export class WorkflowService {
         const documentId = document.id! || getDocumentId(document, false);
 
         const persistedFile = await this.fileService.copyToDestinationAndCreate(
-          { id: documentId, uri: documentPage.uri, provider: documentPage.provider },
+          {
+            id: documentId,
+            uri: documentPage.uri,
+            provider: documentPage.provider,
+            // TODO: Solve once DefaultContextSchema is typed by Typebox
+            fileName: (
+              documentPage as typeof documentPage & {
+                fileName: string;
+              }
+            ).fileName,
+          },
           entityId,
           projectId,
           customerName,
         );
 
         const ballerineFileId = documentPage.ballerineFileId || persistedFile?.id;
+        const mimeType =
+          persistedFile?.mimeType ||
+          mime.getType(persistedFile.fileName || persistedFile.uri || '');
 
-        return { ...documentPage, type: persistedFile?.mimeType, ballerineFileId };
+        return {
+          ...documentPage,
+          type: mimeType,
+          ballerineFileId,
+          fileName: persistedFile?.fileName,
+        };
       }),
     );
   }